# docs/API_SPECIFICATION.yaml
openapi: 3.0.0
info:
  title: ClearDeed API
  version: 1.0.0
  description: End-to-end verified real estate & investment execution platform
  contact:
    name: ClearDeed Advisors LLP
servers:
  - url: https://api.cleardeed.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Development
paths:
...
  /auth/send-otp:
    post:
      summary: Send OTP to mobile number
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mobile_number:
                  type: string
                  example: "+919876543210"
              required:
                - mobile_number
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  request_id:
                    type: string
        '429':
          description: Too many OTP requests (rate limited)
        '400':
          description: Invalid mobile number
  /auth/verify-otp:
    post:
      summary: Verify OTP and login/signup
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mobile_number:
                  type: string
                otp:
                  type: string
              required:
                - mobile_number
                - otp
      responses:
        '200':
          description: OTP verified, user logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  is_new_user:
                    type: boolean
        '401':
          description: Invalid OTP
        '423':
          description: Too many failed attempts, account locked
  /auth/logout:
    post:
      summary: Logout user
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully

  # ======================== PROFILE ========================
...
  /deals:
    post:
      summary: Create deal (admin only)
      tags:
        - Deals
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                buyer_user_id:
                  type: integer
                seller_user_id:
                  type: integer
                property_id:
                  type: integer
                referral_partners:
                  type: array
                  items:
                    type: object
                    properties:
                      referral_partner_id:
                        type: integer
                      side:
                        type: string
                        enum: ['buyer', 'seller']
                      commission_percentage:
                        type: number
              required:
                - buyer_user_id
                - seller_user_id
                - property_id
      responses:
        '201':
          description: Deal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deal'
  /deals/{deal_id}:
    get:
      summary: Get deal details
      tags:
        - Deals
      security:
        - BearerAuth: []
      parameters:
        - name: deal_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deal details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deal'
  /deals/{deal_id}/close:
    post:
      summary: Close deal (admin only)
      tags:
        - Deals
      security:
        - BearerAuth: []
      parameters:
        - name: deal_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deal closed, commissions calculated

  # ======================== COMMISSIONS ========================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        mobile_number:
          type: string
        full_name:
          type: string
        email:
          type: string
        city:
          type: string
        profile_type:
          type: string
          enum: ['buyer', 'seller', 'investor']
        is_verified:
          type: boolean
...
    Deal:
      type: object
      properties:
        id:
          type: integer
        buyer_user_id:
          type: integer
        seller_user_id:
          type: integer
        property_id:
          type: integer
        status:
          type: string
        transaction_value:
          type: number
        created_at:
          type: string
          format: date-time
        referral_mappings:
          type: array
          items:
            type: object
            properties:
              referral_partner_id:
                type: integer
              side:
                type: string
              commission_percentage:
                type: number
...